requirements.txt
black==21.10b0
certifi==2020.12.5
click==8.0.1
et-xmlfile==1.1.0
Flask==2.0.1
Flask-MySQLdb==0.2.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mypy-extensions==0.4.3
mysqlclient==2.0.3
openpyxl==3.0.7
passlib==1.7.4
pathspec==0.9.0
platformdirs==2.4.0
python-dotenv==0.19.2
regex==2021.11.10
tomli==1.2.2
typing_extensions==4.0.0
Werkzeug==2.0.1
WTForms==3.0.0
